Trait browser_window_core::application::ApplicationExt

pub trait ApplicationExt: Copy {
    fn assert_correct_thread(&self);

    fn dispatch(
        &self, 
        work: unsafe fn(_: ApplicationImpl, _: *mut ()), 
        data: *mut ()
    ) -> bool;

    fn exit(&self, exit_code: i32);

    fn exit_threadsafe(&self, exit_code: i32);

    fn initialize(
        argc: c_int, 
        argv: *mut *mut c_char, 
        settings: &ApplicationSettings
    ) -> ApplicationImpl;

    fn run(
        &self, 
        on_ready: unsafe fn(_: ApplicationImpl, _: *mut ()), 
        data: *mut ()
    ) -> i32;

    fn finish(&self) { ... }
}

Required methods

fn assert_correct_thread(&self)

Asserts if not on the GUI thread

fn dispatch(
&self,
work: unsafe fn(_: ApplicationImpl, _: *mut ()),
data: *mut ()
) -> bool

Dispatches work to be executed on the GUI thread.

fn exit(&self, exit_code: i32)

Causes the main loop to exit and lets it return the given code.

fn exit_threadsafe(&self, exit_code: i32)

Same as exit, but is thread-safe.

fn initialize(
argc: c_int,
argv: *mut *mut c_char,
settings: &ApplicationSettings
) -> ApplicationImpl

fn run(
&self,
on_ready: unsafe fn(_: ApplicationImpl, _: *mut ()),
data: *mut ()
) -> i32

Runs the main loop. This blocks until the application is exitting.

Loading content...
Provided methods

fn finish(&self)

Shuts down all application processes and performs necessary clean-up code.

Loading content...